requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
python-dotenv
pandas
openpyxl
Pillow
faker
django-cors-headers
djangorestframework
djangorestframework-simplejwt
djangorestframework-stubs
typing